backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework-simplejwt==5.4.0
django-celery-beat==2.7.0
django-celery-results==2.5.1

psycopg2-binary==2.9.10
redis==5.2.1
celery==5.4.0

boto3==1.35.81
django-storages==1.14.4

Pillow==11.0.0
python-magic==0.4.27
python-decouple==3.8

gunicorn==23.0.0
whitenoise==6.8.2

drf-spectacular==0.28.0
